Moved helpers from DatabaseSavepointObjectMapper to DatabaseSavepointsRepository
